Added date display between messages and fixed background of EmojiBottomSheetFragment

+++ app/src/main/java/com/shudss00/gigachat/data/source/remote/model/MessageDto.kt
@@ -18,5 +18,5 @@ data class MessageDto(
     @SerialName("reactions")
     val reactions: List<ReactionDto>,
     @SerialName("timestamp")
-    val timestamp: Int
+    val timestamp: Long
 )
